Move I2C Pot impls from 'hardware' mod within controls.rs to devices/i2c.rs
