+++ README.md
@@ -2,9 +2,10 @@
 
 Digital synthesizer project for personal entertainment and education.
 
+
 ## Development
 
-TODOs:
+### TODOs
 
 - [ ] Generalize stream pattern to N-channel audio
     - [x] Implement `StereoConsumer`
@@ -32,3 +33,13 @@ TODOs:
       preferable to code littered with `<'a>` explicit lifetimes
 - [ ] Implement some form of CLI for `psynth-play` such that doing new things doesn't always
   involve modifying the `bin/main.rs` and recompiling
+
+
+### Dependencies
+
+On top of those listed in the cargo manifest, the following system packages are required (exact
+package name will depend on your distro):
+
+- `pkg-config`
+- `libzmq` ≥ 4.1
+- `alsa` (see `cpal` docs for more details)
